Fix wither skull lag
